ResponseEntity

+++ ResponseEntity/demo/src/main/java/com/responceentity/demo/Services/UserServicesImp.java
@@ -1,11 +1,13 @@
 package com.responceentity.demo.Services;
 
-import java.beans.Beans;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.responceentity.demo.Dto.DtoUser;
@@ -19,73 +21,70 @@ public class UserServicesImp implements UserServices {
     @Autowired
     UserRepository userRepository;
 
-    // dependacy injeaction in  Constructor
-    public UserServicesImp(UserRepository userRepository ){
-        this.userRepository=userRepository;
+    // dependacy injeaction in Constructor
+    public UserServicesImp(UserRepository userRepository) {
+        this.userRepository = userRepository;
     }
-
 
     @Override
     public ModelUser saveuser(ModelUser modelUser) {
-        
-        DtoUser dtoUser=new DtoUser();
+
+        DtoUser dtoUser = new DtoUser();
         BeanUtils.copyProperties(modelUser, dtoUser);
         userRepository.save(dtoUser);
 
         return modelUser;
     }
 
-
     @Override
     public List<ModelUser> getalluser() {
-       List<ModelUser>modelUsers=new ArrayList<>();
-       List<DtoUser>dtoUserDB=userRepository.findAll();
-       for (DtoUser dtoUser : dtoUserDB) {
-            ModelUser modelUser=new ModelUser();
+        List<ModelUser> modelUsers = new ArrayList<>();
+        List<DtoUser> dtoUserDB = userRepository.findAll();
+        for (DtoUser dtoUser : dtoUserDB) {
+            ModelUser modelUser = new ModelUser();
             BeanUtils.copyProperties(dtoUser, modelUser);
 
             modelUsers.add(modelUser);
-       }
-
-
+        }
 
         return modelUsers;
     }
 
-
     @Override
     public ModelUser getIduser(int id) {
-        DtoUser dtoUser=userRepository.findById(id).get();
-        ModelUser modelUser=new ModelUser();
+        DtoUser dtoUser = userRepository.findById(id).get();
+        ModelUser modelUser = new ModelUser();
         BeanUtils.copyProperties(dtoUser, modelUser);
         return modelUser;
     }
 
-
     @Override
-    public String updateuser(int id, ModelUser modelUser) {
+    public  ResponseEntity<Object> updateuser(int id, ModelUser modelUser) {
 
-        DtoUser dtoUser=userRepository.findById(id).get();
+        DtoUser dtoUser = userRepository.findById(id).get();
 
-     
         dtoUser.setName(modelUser.getName());
 
         userRepository.save(dtoUser);
+        
+        BeanUtils.copyProperties(dtoUser ,modelUser);
 
-       return "User Update Succefully";
+        return ResponseEntity.ok().body("User with ID " + modelUser + " Update successfully.");
     }
-
 
     @Override
-    public String deleteUserById(int id ) {
-        DtoUser dtoUser=userRepository.findById(id).get();
-        userRepository.delete(dtoUser);
-        return " User Is  Deleted ";
+    public ResponseEntity<Object> deleteUserById(int id) {
+        Optional<DtoUser> optionalDtoUser = userRepository.findById(id); // Retrieve Optional
 
-        
+        if (optionalDtoUser.isPresent()) { // Check if user exists
+            DtoUser dtoUser = optionalDtoUser.get(); // Get the user object
+            userRepository.delete(dtoUser); // Delete the user
+            return ResponseEntity.ok().body("User with ID " + id + " deleted successfully.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with ID " + id + " not found.");
+        }
     }
-    
 
-    // delete Service  Make
    
+
 }
